Various Upgrades

+++ images/FontPrint.h
@@ -2,6 +2,6 @@
 #define FONTPRINT_H
 #include <stdint.h>
 
-void printText(char* text, uint8_t x_pos, uint8_t y_pos, uint16_t bg_color, uint16_t text_color);
+void printText(char* text, uint16_t x_pos, uint16_t y_pos, uint16_t bg_color, uint16_t text_color);
 
 #endif
